- Refined venn diagram to include a "sample space box", when user enters input outside of diagram it adds to overall population
- Refined bounds, accomodates up to 2 circles only as of now

- Added features to legend that includes input display, and changeable
title that is reflected on the diagram.

- Changed input feature to only accomodate one input that can be changed
as many times as possible, only one text field is present per
circle/area that shows input

- Added statistical breakdown button that shows a textfield regarding
statistical info about the venn diagram.

**NOTE** The add button should be removed for now as bounds and legend
to not apply for more then 2 circles, this feature will be added later.
The add button is still present as of this build though.